Address a bunch of things that won't compile in Scala 3 anymore.
Mostly, procedure syntax and calling methods with `()`.

+++ editor/src/main/scala/com/funlabyrinthe/editor/pickling/Pickle.scala
@@ -71,6 +71,6 @@ object NumberPickle {
   def unapply(pickle: NumberPickle): Some[Double] = Some(pickle match {
     case FloatPickle(v) => v
     case DoublePickle(v) => v
-    case IntegerPickle(v) => v
+    case IntegerPickle(v) => v.toDouble
   })
 }
